+++ src/shared/sideBar/SideBar.stories.tsx
@@ -9,8 +9,7 @@ import PlusSquare from '@/../public/assets/components/PlusSquare';
 import SearchOutline from '@/../public/assets/components/SearchOutline';
 import TrendingUp from '@/../public/assets/components/TrendingUp';
 import { SideBar } from '@/shared/sideBar/SideBar';
-import { Typography } from '@/shared/ui';
-import { Button } from '@/shared/ui';
+import { Button, Typography } from '@/shared/ui';
 import { StoryProps } from '@storybook/blocks';
 import { Meta, StoryFn } from '@storybook/react';
 import clsx from 'clsx';
