+++ src/shared/ui/alert/ui/Alert.tsx
@@ -1,11 +1,9 @@
 'use client';
 
 import CloseOutline from '@/../public/assets/components/CloseOutline';
-import { Typography } from '@/shared/ui';
-import { Button } from '@/shared/ui';
+import { Button, Typography } from '@/shared/ui';
 import { alertSelectors } from '@/shared/ui/alert/model/alertSlice';
-import { useActions } from '@/shared/utils';
-import { useAppSelector } from '@/shared/utils';
+import { useActions, useAppSelector } from '@/shared/utils';
 import { clsx } from 'clsx';
 
 import s from './Alert.module.scss';
