+++ src/shared/ui/datePicker/lib/helpers/getInputValueDate.ts
@@ -1,5 +1,4 @@
-import { RangeDate } from '@/shared/ui';
-import { getInputValueFromDate } from '@/shared/ui/datePicker/lib/utils';
+import { RangeDate, getInputValueFromDate } from '@/shared/ui';
 
 export const getInputValueDate = (value: RangeDate): string => {
   return `${getInputValueFromDate(value.startDate)} - ${getInputValueFromDate(value.endDate)}`;
