+++ src/shared/ui/datePicker/lib/helpers/updateValueOnPopupCloseAction.ts
@@ -1,6 +1,4 @@
-import { RangeDate } from '@/shared/ui';
-import { getInputValueDate } from '@/shared/ui/datePicker/lib/helpers/getInputValueDate';
-import { getDateFromInputValue, isInRange } from '@/shared/ui/datePicker/lib/utils';
+import { RangeDate, getDateFromInputValue, getInputValueDate, isInRange } from '@/shared/ui';
 
 interface Props {
   inputValue: string;
