+++ src/shared/ui/datePicker/lib/hooks/index.ts
@@ -1,0 +1,2 @@
+export * from './useInputValueRange';
+export * from './useShowPopup';
