+++ src/shared/ui/index.ts
@@ -3,14 +3,18 @@ export * from './backBtn/BackBtn';
 export * from './button/Button';
 export * from './card/Card';
 export * from './checkbox/Checkbox';
+export * from './datePicker/lib';
 export * from './datePicker/ui/DatePicker';
+export * from './datePicker/ui/DatePickerPopupContent';
 export * from './forms/FormCheckbox';
 export * from './forms/FormTextfield';
 export * from './input/Input';
 export * from './mobileMenu/MobileMenu';
 export * from './modal/Modal';
+export * from './pagination/lib/paginationLogic';
 export * from './pagination/ui/PaginationWithSelect';
 export * from './pagination/ui/paginationInitial/Pagination';
+export * from './pagination/ui/paginationInitial/pageLink/PageLink';
 export * from './privacy/Privacy';
 export * from './radio/Radio';
 export * from './select/Select';
