+++ src/shared/ui/pagination/ui/paginationInitial/Pagination.tsx
@@ -1,7 +1,6 @@
 import SvgArrowIosBack from '@/../public/assets/components/ArrowIosBack';
 import SvgArrowIosForward from '@/../public/assets/components/ArrowIosForward';
-import { usePagination } from '@/shared/ui/pagination/lib/paginationLogic';
-import { PageLink } from '@/shared/ui/pagination/ui/paginationInitial/pageLink/PageLink';
+import { PageLink, usePagination } from '@/shared/ui';
 
 import s from './Pagination.module.scss';
